requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-pagination==2.3.2
Django==2.0.7
django-appconf==1.0.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==7.1.0
pytz==2018.5
rjsmin==1.1.0
six==1.13.0
